+++ example/lib/tablevalues_page.dart
@@ -1,5 +1,3 @@
-import 'dart:math';
-
 import 'package:flutter/material.dart';
 import 'package:graphic_representation/graphic_representation.dart';
 
@@ -22,7 +20,7 @@ class TableValuePage extends StatelessWidget {
           size: Size(MediaQuery.of(context).size.width,
               MediaQuery.of(context).size.height * 0.2),
           functionName: 'x² + 1/x',
-          f: (x) => x*x + 1/x,
+          f: (x) => x * x + 1 / x,
           numsX: [-2, -1, 0, 1, 2],
           fontSize: 12,
         )